prepare bumping bone limit

+++ src/main/java/com/github/worldsender/mcanm/client/mcanmmodel/stored/RawDataV1.java
@@ -74,7 +74,7 @@ public class RawDataV1 implements IVersionSpecificData, ISkeletonVisitable {
                 throw new ModelFormatException(
                         String.format("ParentIndex (%d) has to be smaller than nbrBones (%d).", parentIndex, nbrBones));
             }
-            bone.parent = (byte) parentIndex;
+            bone.parent = parentIndex;
         }
     }
