+++ src/main/java/com/github/worldsender/mcanm/common/skeleton/stored/RawDataV1.java
@@ -41,7 +41,7 @@ public class RawDataV1 implements IVersionSpecificData {
                 throw new ModelFormatException(
                         String.format("ParentIndex (%d) has to be smaller than nbrBones (%d).", parentIndex, nbrBones));
             }
-            bone.parent = (byte) parentIndex;
+            bone.parent = parentIndex;
         }
     }
